Added test_resources.py
